# Main site optimizations