--- a/static/css/_pages-members.css
+++ b/static/css/_pages-members.css
@@ -1,0 +1,4 @@
+.member-profile-img {
+      width: 30px;
+      height: 30px;
+}