… open_humans/templates/partials/activity-panel.html
… private_sharing/migrations/0012_auto_20190115_2014.py
… public_data/templates/public_data/home.html
… static/images/your-data-plus-you-sketch.jpg